requirements.txt
customtkinter==5.1.2
Pillow==9.5.0
diffusers==0.14.0
transformers==4.28.0
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1